+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/assembly/KafkaConnectAssemblyOperator.java
@@ -163,6 +163,8 @@ public class KafkaConnectAssemblyOperator extends AbstractConnectOperator<Kubern
 
         connectServiceAccount(reconciliation, namespace, KafkaConnectResources.serviceAccountName(connect.getCluster()), connect)
                 .compose(i -> connectInitClusterRoleBinding(reconciliation, initCrbName, initCrb))
+                .compose(i -> connectRole(reconciliation, namespace, connect))
+                .compose(i -> connectRoleBinding(reconciliation, namespace, connect))
                 .compose(i -> connectNetworkPolicy(reconciliation, namespace, connect, isUseResources(kafkaConnect)))
                 .compose(i -> manualRollingUpdate(reconciliation, connect))
                 .compose(i -> podSetOperations.getAsync(reconciliation.namespace(), connect.getComponentName()))